Revert "feat: logic moved from controller to service"
This reverts commit 73fe63786dce34c3e6689ef502a40b68fa7528fb.
